Split statements tests out
