+++ .github/CODEOWNERS
@@ -101,6 +101,7 @@ assets/               @DataDog/agent-integrations
 **/postgres/statement_samples.py             @DataDog/agent-integrations @DataDog/database-monitoring
 **/mysql/statements.py                       @DataDog/agent-integrations @DataDog/database-monitoring
 **/mysql/statement_samples.py                @DataDog/agent-integrations @DataDog/database-monitoring
+**/mysql/tests/test_statements.py            @DataDog/agent-integrations @DataDog/database-monitoring
 
 # Checks base
 /datadog_checks_base/                                          @DataDog/agent-core @DataDog/agent-integrations
